Minor Character Details UI change
Centered the bottom buttons, Bugs found: Feats table causes type and
source to be placed under the name of the feat without visual
separation. Due to the quantity of Feats, feats-all hangs for about 5-15
seconds on first load and 2-5 seconds on loads their after. Recommend
switching from table to list similar to character combat. Might improve
performance if we split the loading by source and append as they load,
to display content to the user and hide the load times. Future feedback.